services/github-mcp-server/requirements.txt
mcp>=1.0.0
PyGithub>=2.1.1
pydantic>=2.5.0
python-dotenv>=1.0.0
cryptography>=41.0.0
pyjwt>=2.8.0
